Refactor output handling in Tunny.Handler.OutputLoop.cs and Tunny.Solver.Solver.cs
